Changed All

+++ src/TechStation.Domain/Entities/UserRole.cs
@@ -3,9 +3,9 @@ using TechStation.Domain.Enums;
 
 namespace TechStation.Domain.Entities;
 
-public class UserRole : Auditable
-{
-    public int UserId { get; set; }
-    public User User { get; set; }
-    public Role Role { get; set; }
-}
+    public class UserRole : Auditable
+    {
+        public long UserId { get; set; }
+        public User User { get; set; }
+        public Role Role { get; set; }
+    }
